Added In-Game State And LOTS of additions / changes
Converted To A Level Based System which incorporates the old Tile[][].
Level Selector Menu Is Now Functional And Renders Differently. Can
switch tabs between campaign mode and custom map mode in the level
selector. Can change pages using keys 3 and 4 on the Level Selector.
Added A Money Handler To Deal With The Money And Added Background
Settings And renders to levels. Many other tweaks and additions